Move and cleanup ldap search UI files
